pyproject.toml
[tool.poetry]
name = "jrachkatop"
version = "0.1.0"
description = ""
authors = ["kuptsovsasha <kuptsovsasha@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
django-environ = "0.9.0"
psycopg2-binary = "2.9.4"
django-rq = "^2.6.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
django-filter = "^22.1"
django-factory_boy = "^1.0.0"
base = "^1.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
